Add clickable receipts in analytics drill-down with edit modal
- Charts drill-down receipts are now clickable buttons that open the
  same EditExpenseForm modal used on the main page
- Pass cards/currentUser/allUsers into AnalyticsPanel for the form
- After save or delete, the drill-down list refreshes automatically
- Also includes earlier changes: category/date filter on list endpoint,
  merchant grouping fix, float qty support, password show/hide toggle

+++ backend/models.py
@@ -6,7 +6,7 @@ from typing import Optional
 
 class ExpenseItem(BaseModel):
     name: str
-    qty: int = 1
+    qty: float = 1
     unit_price: float | None = None
     amount: float
